Rename `FeedParser` to `XmlFeedParser`
